Pull out Http utils
